% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_utils.R
\name{content_text}
\alias{content_text}
\title{Extracts text from html object given css selector}
\usage{
content_text(webpage, css_sel)
}
\arguments{
\item{webpage}{xml2 object read_html}

\item{css_sel}{css selector or xpath pointing to text that supposed to
be downloaded}
}
\value{
A character.
}
\description{
Extracts text from html object given css selector
}
