% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_utils.R
\name{creation_date}
\alias{creation_date}
\title{Extracts creation date from oxl flat rental offer}
\usage{
creation_date(webpage)
}
\arguments{
\item{webpage}{xml2 object read_html}
}
\value{
date
}
\description{
Extracts creation date from oxl flat rental offer
}
