% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_utils.R
\name{discription}
\alias{discription}
\title{Extracts discription from oxl flat rental offer}
\usage{
discription(webpage)
}
\arguments{
\item{webpage}{xml2 object read_html}
}
\value{
character
}
\description{
Extracts discription from oxl flat rental offer
}
