% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_utils.R
\name{district}
\alias{district}
\title{Extracts name of district from oxl flat rental offer}
\usage{
district(webpage)
}
\arguments{
\item{webpage}{xml2 object read_html}
}
\value{
character
}
\description{
Extracts name of district from oxl flat rental offer
}
