% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_utils.R
\name{extract_table}
\alias{extract_table}
\title{Extracts table with detail from oxl flat rental offer}
\usage{
extract_table(webpage)
}
\arguments{
\item{webpage}{xml2 object read_html}
}
\value{
tibble
}
\description{
Extracts table with detail from oxl flat rental offer
}
