% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_olx.R
\name{get_offers}
\alias{get_offers}
\title{get_offers scraps flat renting offers posted on OLX.pl}
\usage{

  get_offers(url_start = "https://www.olx.pl/nieruchomosci/mieszkania/wynajem/wroclaw")
}
\arguments{
\item{url_start}{starting url, default one is for site with offers from
Wrocław}
}
\value{
A tibble containing data about offers
}
\description{
It starts at given url, determines how many sites with links to offers are
ther, scraps links to offers and than by crawling from link to link
downloads date about offers.
}
\details{
Some links points to offers on sites other than OLX. Those links are ommited.
}
