% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_utils.R
\name{h_rent}
\alias{h_rent}
\title{Parses value of rent extracted from table with deatils}
\usage{
h_rent(value_PLN)
}
\arguments{
\item{value_PLN}{character}
}
\value{
double
}
\description{
Parses value of rent extracted from table with deatils
}
