% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_utils.R
\name{h_size}
\alias{h_size}
\title{Parses size of flat extracted from table with deatils}
\usage{
h_size(size_m)
}
\arguments{
\item{size_m}{character}
}
\value{
double
}
\description{
Parses size of flat extracted from table with deatils
}
