% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_utils.R
\name{if_empty}
\alias{if_empty}
\title{Determines whether object is empty, if so assigns NA to it.}
\usage{
if_empty(obj)
}
\arguments{
\item{obj}{atomic vector, list, matrix or data frame}
}
\value{
passed object or NA if object is empty
}
\description{
Determines whether object is empty, if so assigns NA to it.
}
