% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_utils.R
\name{name}
\alias{name}
\title{Extracts name of oxl flat rental offer}
\usage{
name(webpage)
}
\arguments{
\item{webpage}{xml2 object read_html}
}
\value{
character
}
\description{
Extracts name of oxl flat rental offer
}
