% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_olx.R
\name{num_of_offer_sites}
\alias{num_of_offer_sites}
\title{num_of_offer_sites allows to determin how many sites listings offers}
\usage{
num_of_offer_sites(webpage)
}
\arguments{
\item{webpage}{- rvest object read_html}
}
\value{
number of pages listings offers
}
\description{
num_of_offer_sites allows to determin how many sites listings offers
}
