% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_olx.R
\name{olx_tab}
\alias{olx_tab}
\title{parsing and formatting of table containg details about olx offer}
\usage{
olx_tab(webpage, css_sel = ".details")
}
\arguments{
\item{webpage}{- rvest object read_html}

\item{css_sel}{- css selector or xpath pointing to text that supposed to
be downloaded}
}
\value{
- tibble containig detail of offer
}
\description{
parsing and formatting of table containg details about olx offer
}
