% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_utils.R
\name{price}
\alias{price}
\title{Extracts price from oxl flat rental offer}
\usage{
price(webpage)
}
\arguments{
\item{webpage}{xml2 object read_html}
}
\value{
double
}
\description{
Extracts price from oxl flat rental offer
}
