% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_olx.R
\name{scrap_offer}
\alias{scrap_offer}
\title{scrap_offer scraps information from flat renting offer on OLX.PL}
\usage{
scrap_offer(url)
}
\arguments{
\item{url}{- url of offer that you want to scrap}
}
\value{
A tibble containing data about offer
}
\description{
scrap_offer scraps information from flat renting offer on OLX.PL
}
