% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_utils.R
\name{user}
\alias{user}
\title{Extracts user name from oxl flat rental offer}
\usage{
user(webpage)
}
\arguments{
\item{webpage}{xml2 object read_html}
}
\value{
character
}
\description{
Extracts user name from oxl flat rental offer
}
