% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_utils.R
\name{views}
\alias{views}
\title{Extracts number of views from oxl flat rental offer}
\usage{
views(webpage)
}
\arguments{
\item{webpage}{xml2 object read_html}
}
\value{
integer
}
\description{
Extracts number of views from oxl flat rental offer
}
