% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_olx.R
\name{get_offers}
\alias{get_offers}
\title{Scraps flat renting offers posted on OLX.pl available at given moment.}
\usage{
get_offers(links = NA)
}
\arguments{
\item{links}{character vector contatining links to offers}
}
\value{
A tibble containing data about offers
}
\description{
If links = NA then function scraps links to offers and than by
  crawling from link to link downloads date about offers. Some links points
  to offers on sites other than OLX. Those links are ommited.
}
